pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "goygoy"
version = "0.1.0"
description = "Real-time speech-to-speech voice chat"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "cerebras-cloud-sdk>=1.0.0",
    "fastapi>=0.115.0",
    "numpy",
    "python-dotenv",
    "pyyaml",
    "snac",
    "soundfile",
    "torch",
    "transformers",
    "uvicorn[standard]>=0.34.0",
    "vllm>=0.8.0",
    "websockets>=14.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "pytest-asyncio",
]

[project.scripts]
goygoy = "server.app:main"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.hatch.build.targets.wheel]
packages = ["goygoy", "server"]
